requirements-dev.txt
-r requirements.txt

isort
black
pytest-django